refactor(client): mv rt-{intents,interop}
