+++ src/client/src/rt-interop/fdc3/adapter/types.ts
