+++ src/client/src/rt-interop/intents/defaultWindowConfig.ts
