+++ src/client/src/rt-interop/intents/responseMapper.ts
